adding vuex modules, and updating the endpoint for marketcap
