+++ package.json
@@ -19,7 +19,8 @@
     "vue": "^2.5.2",
     "vue-popperjs": "^1.2.6",
     "vue-router": "^3.0.1",
-    "vuex": "^3.0.1"
+    "vuex": "^3.0.1",
+    "vuex-loading": "^0.3.0"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
