+++ src/store/getters/index.js
@@ -1,0 +1,13 @@
+/**
+ * Exporting Getters
+ */
+export default {
+  /**
+   * Template
+   * @param {{}} state
+   * @return {Function}
+   */
+  template: function (state) {
+    return state
+  }
+}
