+++ src/store/modules/_latestBlock/index.js
@@ -1,0 +1,14 @@
+/**
+ * Exporting Vuex Module
+ */
+export default {
+  namespaced: true,
+  state: {
+    block: null
+  },
+  mutations: {
+    setBlock (state, block) {
+      return Object.assign(state, { block })
+    }
+  }
+}
