+++ src/store/modules/_marketStats/index.js
@@ -1,0 +1,67 @@
+/**
+ * Importing API
+ */
+import { createActionHelpers } from 'vuex-loading'
+
+/**
+ * Setting up start/end Loading helper methods
+ */
+const { startLoading, endLoading } = createActionHelpers({
+  moduleName: 'loading'
+})
+
+/**
+ * Exporting Vuex Module
+ */
+export default {
+  namespaced: true,
+  state: {
+    priceChf: null,
+    marketCapChf: null,
+    priceBtc: null
+  },
+  mutations: {
+    setMarketStats (state, stats) {
+      return Object.assign(state, stats)
+    }
+  },
+  actions: {
+    /**
+     * get market stats of the overall blockchain
+     * @param {Object} state
+     * @param {Function} commit
+     * @return {Promise<void>}
+     */
+    async get ({ state, commit }) {
+      /**
+       * Start vuex-loading
+       */
+      startLoading('_marketStats/get')
+
+      /**
+       * Pulling market stats
+       * @type {Response}
+       */
+      const market = await fetch('https://api.coinmarketcap.com/v1/ticker/aeternity/?convert=CHF');
+
+      /**
+       * Renaming variables
+       */
+      const {
+        price_btc: priceBtc,
+        market_cap_chf: marketCapChf,
+        price_chf: priceChf
+      } = (await market.json())[0]
+
+      /**
+       * end vuex-loading
+       */
+      endLoading('_marketStats/get')
+
+      /**
+       * committing to the state
+       */
+      return commit('setMarketStats', { priceBtc, marketCapChf, priceChf })
+    }
+  }
+}
