+++ src/store/modules/_recentBlocks/index.js
@@ -1,0 +1,14 @@
+/**
+ * Exporting Vuex Module
+ */
+export default {
+  namespaced: true,
+  state: {
+    recent: []
+  },
+  mutations: {
+    setRecent (state, recent) {
+      return Object.assign(state, { recent })
+    }
+  }
+}
