+++ src/store/modules/accounts/getters.js
@@ -1,0 +1,13 @@
+/**
+ * Exporting Getters
+ */
+export default {
+  /**
+   * Return block array
+   * @param {Object} state
+   * @return {Array}
+   */
+  blocks: function (state) {
+    return state.blocks
+  }
+}
