+++ src/store/modules/accounts/mutations.js
@@ -1,0 +1,14 @@
+/**
+ * Exporting Mutations
+ */
+export default {
+  /**
+   * Set blocks
+   * @param {Object} state
+   * @param {Boolean} blocks
+   * @return {Object}
+   */
+  setBlocks: function (state, blocks) {
+    return Object.assign(state, { blocks })
+  }
+}
