+++ src/store/modules/blocks/state.js
@@ -1,0 +1,6 @@
+/**
+ * Exporting State
+ */
+export default {
+  blocks: []
+}
