+++ src/store/modules/transactions/getters.js
@@ -1,0 +1,13 @@
+/**
+ * Exporting Getters
+ */
+export default {
+  /**
+   * Return transactions array
+   * @param {Object} state
+   * @return {Array}
+   */
+  transactions: function (state) {
+    return state.transactions
+  }
+}
